nambahin fitur ajax

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -24,13 +25,17 @@ class PostController extends Controller
 		}
 
 	public function getDashboard(){
-		$posts = Post::all();
-   	 return view('dashboard',['posts'=> $posts]);
+		$posts = Post::orderBy('created_at','desc')->get();
+   	   return view('dashboard',['posts'=> $posts]);
    }
 
    public function getDeletePost($post_id)
    {
    	$post = Post::where('id', $post_id)->first();
+   	if (Auth::user() != $post->user) {
+   		# code...
+   		return redirect()->back();
+   	}
    	$post->delete();
    	return redirect()->route('dashboard')->with(['message'=>'sucessfully deleted']);
    }
